User, Auth 및 Controller 클래스 리팩토링 및 기능 추가 - `User` 엔티티 필드 및 생성자 수정 (`userId` → `UserEmail`) - `LoginRequest`, `LoginResponse` 필드 및 메서드 수정 (`isOnline` 추가) - `AuthService` 로그인 로직 개선 (`UserStatus` 연동) - 신규 Controller 클래스 추가 (`UserController`, `LoginController`, `BinaryContentController` 등) - DTO 추가 및 변경 (`UserStatustUpdateRequest`, `UserDeleteRequest` 등) - 예외 메시지 및 일부 로직 개선 (e.g., `UserStatusRepository`) - `application.yml` 리파지토리 설정값 수정

+++ src/main/java/com/sprint/mission/discodeit/service/jcf/basic/BasicUserService.java
@@ -51,9 +51,9 @@ public class BasicUserService implements UserService {
      
      //유저정보
         User user = new User(
+                userCreateRequest.email(),
+                userCreateRequest.rawPassword(),
                 userCreateRequest.username()
-                ,userCreateRequest.email()
-                ,userCreateRequest.rawPassword()
                 ,userCreateRequest.userNickname()
         ); 
      
@@ -118,11 +118,14 @@ public class BasicUserService implements UserService {
         if (!userRepository.existsById(userId)) {
             throw new NoSuchElementException("유저uuid못찾아용"+userId);
         }
+
         UUID profileID = userRepository
                 .findById(userId)
                 .orElseThrow(() -> new NoSuchElementException("유저uuid못찾아용" + userId))
                 .getProfileID();
-         binaryRepository.delete(profileID);
+        if(profileID != null) {
+            binaryRepository.delete(profileID);
+        }
         //이런게 너무번거롭다 싫다
         userStatusRepository.findByUserId(userId);
         userRepository.deleteById(userId);
